delete condition check

+++ app/Http/Livewire/Client/ClientComponent.php
@@ -3,6 +3,9 @@
 namespace App\Http\Livewire\Client;
 
 use App\Models\Client;
+use App\Models\FieldDensityCdot;
+use App\Models\FieldDensityCommercial;
+use App\Models\TemplateOne;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -20,15 +23,24 @@ class ClientComponent extends Component
         $this->dispatchBrowserEvent('show-delete-confirmation');
     }
 
+
     public function deleteData()
     {
-        $data = Client::find($this->delete_id);
-        $data->delete();
-        $this->delete_id = '';
+        $cdot = FieldDensityCdot::where('client_id', $this->delete_id)->get();
+        $commercial = FieldDensityCommercial::where('client_id', $this->delete_id)->get();
+        $temp_one = TemplateOne::where('client_id', $this->delete_id)->get();
 
-        $this->dispatchBrowserEvent('ClientDeleted');
+        if ($cdot->count() > 0 || $commercial->count() > 0 || $temp_one->count() > 0) {
+            $this->dispatchBrowserEvent('clientDeleteError');
+        } else {
+            $data = Client::find($this->delete_id);
+            $data->delete();
 
+            $this->dispatchBrowserEvent('ClientDeleted');
+            $this->resetInputs();
+        }
     }
+
 
     public function render()
     {
